django_rest_api_w_vue/requirements.txt
Django==1.10.6
django-filter==1.1.0
djangorestframework==3.7.1
